refactor core to simplify file structure
