+++ packages/core/src/index.ts
@@ -6,13 +6,12 @@ export * from './actions';
 export * from './database';
 export * from './entities';
 export * from './logger';
+export * from './prompts';
 export * from './roles';
 export * from './runtime';
-export * from './settings';
-export * from './uuid';
-export * from './audioUtils';
-export * from './utils';
 export * from './search';
+export * from './settings';
+export * from './utils';
 
 // Export instrumentation types and service
 export * from './instrumentation/types';
