+++ packages/core/src/prompts.ts
@@ -1,0 +1,103 @@
+export const shouldRespondTemplate = `<task>Decide on behalf of {{agentName}} whether they should respond to the message, ignore it or stop the conversation.</task>
+
+<providers>
+{{providers}}
+</providers>
+
+<instructions>Decide if {{agentName}} should respond to or interact with the conversation.
+If the message is directed at or relevant to {{agentName}}, respond with RESPOND action.
+If a user asks {{agentName}} to be quiet, respond with STOP action.
+If {{agentName}} should ignore the message, respond with IGNORE action.</instructions>
+
+<output>
+Respond using XML format like this:
+<response>
+  <name>{{agentName}}</name>
+  <reasoning>Your reasoning here</reasoning>
+  <action>RESPOND | IGNORE | STOP</action>
+</response>
+
+Your response should ONLY include the <response></response> XML block.
+</output>`;
+
+export const messageHandlerTemplate = `<task>Generate dialog and actions for the character {{agentName}}.</task>  
+
+<providers>
+{{providers}}
+</providers>
+
+These are the available valid actions:
+<actionNames>
+{{actionNames}}
+</actionNames>
+
+<instructions>
+Write a thought and plan for {{agentName}} and decide what actions to take. Also include the providers that {{agentName}} will use to have the right context for responding and acting, if any.
+First, think about what you want to do next and plan your actions. Then, write the next message and include the actions you plan to take.
+</instructions>
+
+<keys>
+"thought" should be a short description of what the agent is thinking about and planning.
+"actions" should be a comma-separated list of the actions {{agentName}} plans to take based on the thought (if none, use IGNORE, if simply responding with text, use REPLY)
+"providers" should be an optional comma-separated list of the providers that {{agentName}} will use to have the right context for responding and acting
+"evaluators" should be an optional comma-separated list of the evaluators that {{agentName}} will use to evaluate the conversation after responding
+"text" should be the text of the next message for {{agentName}} which they will send to the conversation.
+"simple" should be true if the message is a simple response and false if it is a more complex response that requires planning, knowledge or more context to handle or reply to.
+</keys>
+
+<output>
+Respond using XML format like this:
+<response>
+    <thought>Your thought here</thought>
+    <actions>ACTION1,ACTION2</actions>
+    <providers>PROVIDER1,PROVIDER2</providers>
+    <text>Your response text here</text>
+    <simple>true|false</simple>
+</response>
+
+Your response must ONLY include the <response></response> XML block.
+</output>`;
+
+export const postCreationTemplate = `# Task: Create a post in the voice and style and perspective of {{agentName}} @{{twitterUserName}}.
+
+Example task outputs:
+1. A post about the importance of AI in our lives
+<response>
+  <thought>I am thinking about writing a post about the importance of AI in our lives</thought>
+  <post>AI is changing the world and it is important to understand how it works</post>
+  <imagePrompt>A futuristic cityscape with flying cars and people using AI to do things</imagePrompt>
+</response>
+
+2. A post about dogs
+<response>
+  <thought>I am thinking about writing a post about dogs</thought>
+  <post>Dogs are man's best friend and they are loyal and loving</post>
+  <imagePrompt>A dog playing with a ball in a park</imagePrompt>
+</response>
+
+3. A post about finding a new job
+<response>
+  <thought>Getting a job is hard, I bet there's a good tweet in that</thought>
+  <post>Just keep going!</post>
+  <imagePrompt>A person looking at a computer screen with a job search website</imagePrompt>
+</response>
+
+{{providers}}
+
+Write a post that is {{adjective}} about {{topic}} (without mentioning {{topic}} directly), from the perspective of {{agentName}}. Do not add commentary or acknowledge this request, just write the post.
+Your response should be 1, 2, or 3 sentences (choose the length at random).
+Your response should not contain any questions. Brief, concise statements only. The total character count MUST be less than 280. No emojis. Use \\n\\n (double spaces) between statements if there are multiple statements in your response.
+
+Your output should be formatted in XML like this:
+<response>
+  <thought>Your thought here</thought>
+  <post>Your post text here</post>
+  <imagePrompt>Optional image prompt here</imagePrompt>
+</response>
+
+The "post" field should be the post you want to send. Do not including any thinking or internal reflection in the "post" field.
+The "imagePrompt" field is optional and should be a prompt for an image that is relevant to the post. It should be a single sentence that captures the essence of the post. ONLY USE THIS FIELD if it makes sense that the post would benefit from an image.
+The "thought" field should be a short description of what the agent is thinking about before responding, inlcuding a brief justification for the response. Includate an explanation how the post is relevant to the topic but unique and different than other posts.
+Your reponse should ONLY contain the XML block.`;
+
+export const booleanFooter = 'Respond with only a YES or a NO.';
